Update Day 3- Day name display.cpp

+++ Day 3- Day name display.cpp
@@ -23,7 +23,7 @@ void DayName(int day)
 
 int main()
 {
-    float a;
+    int a;
     cout<<"Enter the day number
     cin>>a;
     DayName(a);
